pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "cntencra",email = "cn10cra@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "psycopg[binary,pool] (>=3.2.9,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)"
]

[tool.poetry]
packages = [{ include = "fastapi_backend" }]

[tool.poetry.scripts]
seed-db = "fastapi_backend.seeds.seed:run_seed"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-env = "^1.1.5"

[tool.pytest.ini_options]
env = ["ENV=test"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
